+++ lib/src/provider.dart
@@ -3,16 +3,16 @@ import 'package:flutter/material.dart';
 import 'package:sprinkle/sprinkle.dart';
 
 class Provider extends InheritedWidget {
-  final Supervisor data;
+  final Supervisor? data;
 
   Provider({
-    Key key,
-    Widget child,
+    Key? key,
+    Widget? child,
     this.data,
-  }) : super(key: key, child: child);
+  }) : super(key: key, child: child!);
 
-  static Supervisor of(BuildContext context) {
-    return (context.dependOnInheritedWidgetOfExactType<Provider>()).data;
+  static Supervisor? of(BuildContext context) {
+    return (context.dependOnInheritedWidgetOfExactType<Provider>())!.data;
   }
 
   @override