+++ lib/src/sprinkle.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:sprinkle/sprinkle.dart';
 
 class Sprinkle extends StatelessWidget {
-  final Widget child;
-  final Supervisor supervisor;
+  final Widget? child;
+  final Supervisor? supervisor;
 
   Sprinkle({this.child, this.supervisor});